% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{bearing_2d}
\alias{bearing_2d}
\title{Compute the bearing between two points}
\usage{
bearing_2d(mat)
}
\arguments{
\item{mat}{A 2x2 matrix with X and Y coordinate in columns, and start and end 
point as rows.}
}
\value{
The bearing (angle) in degrees. 0 indicates North.
}
\description{
Computes a 2d bearing (angle) of a line between two points.
}
