% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend_rasters.R
\name{blend_rasters}
\alias{blend_rasters}
\title{Blend two rasters together by weighted mean}
\usage{
blend_rasters(r1, r2, l1, l2, blenddist)
}
\arguments{
\item{r1}{A SpatRaster}

\item{r2}{A second spatRaster}

\item{l1}{A line associated with r1}

\item{l2}{A line associated with r2}

\item{blenddist}{the distance where blending will be made}
}
\value{
a blended raster
}
\description{
The function blends two overlapping rasters together by computing a weighted
mean between the two rasters. Weights are derived by distance from linestrings
associated with each raster.
}
