% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{create_shape}
\alias{create_shape}
\title{Creates the list of lengths and angles for use in strip_*()-functions}
\usage{
create_shape(line, tolerance)
}
\arguments{
\item{line}{a sf linestring object showing the target line.}

\item{tolerance}{The number of transects to create from the line}
}
\value{
a list with three elements: 1. the simplified shape, and its 
associated 2. lengths, and 3. angles.
}
\description{
The function takes an input of a linestring, which is simplified based on
given tolerance. The lengths and angles between nodes are then recorded.
}
