% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transects.R
\name{create_transects}
\alias{create_transects}
\title{Computes transects at linestring nodes}
\usage{
create_transects(sf, window_size = 3, width)
}
\arguments{
\item{sf}{An sf linestring}

\item{window_size}{how many nodes to consider when making a transect}

\item{width}{The width of the transect in the same units as the CRS of linestring.}
}
\value{
an sf LINESTRING object
}
\description{
The function computes transects of a specified width from an sf linestring.
The function is a translation from Jo Wood's code in 
https://observablehq.com/@jwolondon/tdmc2024_day05.
}
