% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{A vector of degrees to convert}
}
\value{
Vector of degrees
}
\description{
Convert degrees to radians
}
