% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_line}
\alias{example_line}
\title{A line used for examples}
\format{
An 'sf' object
}
\usage{
example_line
}
\description{
An sf object with linestring geometries to use in the examples of some 
200km in length..
}
\keyword{datasets}
