% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{move_coords}
\alias{move_coords}
\title{Compute displacement in X and Y coordinates}
\usage{
move_coords(bear, dist)
}
\arguments{
\item{bear}{Angle of bearing}

\item{dist}{distance}
}
\value{
A vector of two values: displacement in X and Y coordinate.
}
\description{
Compute displacement in X and Y coordinates based on bearing angle and 
distance along the bearing angle.
}
