% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert radians to degrees}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{A vector of radians to convert}
}
\value{
Vector of degrees
}
\description{
Convert radians to degrees
}
