% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_raster.R
\name{rotate_raster}
\alias{rotate_raster}
\title{Rotate a raster around a specified point}
\usage{
rotate_raster(r, angle, point = NULL)
}
\arguments{
\item{r}{A SpatRaster to be rotated}

\item{angle}{Angle in degrees, with positive angle in clock-wise direction}

\item{point}{A sf POINT indicating the point around which rotation should
be done. if NULL, centroid of r will be used.}
}
\value{
a rotated SpatRaster
}
\description{
The function rotates a raster around a specified point.Inspired by Damien
Caillaud's answer in Stackoverflow: https://gis.stackexchange.com/a/479974.
}
