% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rotation_matrix}
\alias{rotation_matrix}
\title{Create rotation matrix for rotation of coordinates}
\usage{
rotation_matrix(a)
}
\arguments{
\item{a}{Angle in degrees, with positive angles in clockwise direction.}
}
\value{
a rotated SpatRaster
}
\description{
Create rotation matrix for rotation of coordinates
}
