% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_raster.R
\name{strip_raster}
\alias{strip_raster}
\title{Create a strip map with raster data to a specified shape}
\usage{
strip_raster(line, r, tolerance, buffer_size, shape = NULL, verbose = FALSE)
}
\arguments{
\item{line}{An sf linestring}

\item{r}{A SpatRaster}

\item{tolerance}{Tolerance for the Douglas-Peucker simplification algorithm.
See details.}

\item{buffer_size}{Buffer size applied to the linestring and used to mask 
the raster.}

\item{shape}{A list of angles and lengths for each segment of the shape. 
Potentially obtained using create_shape(). If left NULL, the shape is a
straight line pointing upward (angle = 0). See details.}

\item{verbose}{Whether or not to print progress information.}
}
\value{
a list with two elements: the rotated linestring and raster.
}
\description{
The function creates a strip of a specified shape, based on a linestring to 
be bent to shape. The raster is rotated, shifted and blended in sequence
from the start of the line. The rasters and linestrings are placed so that
they approximately follow the target shape.
}
