% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_vector.R
\name{strip_vector}
\alias{strip_vector}
\title{Create a strip map with vector data to a specified shape. EXPERIMENTAL}
\usage{
strip_vector(line, v, tolerance, buffer_size, shape, verbose = FALSE)
}
\arguments{
\item{line}{An sf linestring}

\item{v}{sf object with the vectors to be strip'd.}

\item{tolerance}{Tolerance for the Douglas-Peucker simplification algorithm.
See details.}

\item{buffer_size}{Buffer size applied to the linestring and used to mask 
the vectors.}

\item{shape}{A list of angles and lengths for each segment of the shape. 
Potentially obtained using create_shape(). If left NULL, the shape is a
straight line pointing upward (angle = 0). See details.}

\item{verbose}{Whether or not to print progress information.}
}
\value{
a list with two elements: the rotated linestring and vectors.
}
\description{
The function creates a strip of a specified shape, based on a linestring to 
be bent to shape. The vectors are rotated, shifted and blended in sequence
from the start of the line. The vectors and linestrings are placed so that
they approximately follow the target shape.
}
