% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{target_shapes}
\alias{target_shapes}
\title{Shapes to use for the examples}
\format{
An 'sf' object
}
\usage{
target_shapes
}
\description{
Contains a few shapes to show a tutorial of forcing the strip to take a 
certain shape.
}
\keyword{datasets}
