% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripR-package.R
\name{watmisc}
\alias{watmisc}
\title{stripR: An R package to support the creation of strip maps}
\description{
An R package dedicated to the processing of spatial data to support
creation of stip maps.
}
